+++ extensions/gitlab/src/components/reviews.tsx
@@ -51,6 +51,7 @@ export function ReviewList() {
 export function useMyReviews(
   project?: Project | undefined,
   labels: string[] | undefined = undefined,
+  params?: Record<string, boolean | number | string>,
 ): {
   mrs: MergeRequest[] | undefined;
   isLoading: boolean;
@@ -59,7 +60,7 @@ export function useMyReviews(
 } {
   const [mrs, setMrs] = useState<MergeRequest[]>();
   const { data, isLoading, error, performRefetch } = useCache<MergeRequest[] | undefined>(
-    `myreviews_${labels ? labels.join(",") : "[]"}`,
+    `myreviews_${labels ? labels.join(",") : "[]"}_${params ? JSON.stringify(params) : ""}`,
     async (): Promise<MergeRequest[] | undefined> => {
       const user = await gitlab.getMyself();
       return await gitlab.getMergeRequests({
@@ -68,10 +69,11 @@ export function useMyReviews(
         in: "title",
         scope: "all",
         ...(labels && { labels }),
+        ...(params || {}),
       });
     },
     {
-      deps: [labels],
+      deps: [labels, JSON.stringify(params)],
       secondsToRefetch: 5,
       secondsToInvalid: daysInSeconds(7),
     },